requirements.txt
runpod==1.6.2
requests
numpy
pillow
tqdm
imageio
trimesh
open3d==0.18.0
pygltflib
scipy
rembg==2.0.57
onnxruntime-gpu==1.18.1
# CLI helpers
click